backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
spacy==3.7.4
transformers==4.42.4
pillow==10.2.0
opencv-python==4.9.0.80
scikit-learn==1.5.0
python-multipart==0.0.9
numpy==1.26.4
pytesseract
Pillow
easyocr
opencv-python

